calcular balance gastos

+++ Misc/spring-boot-gastos-compartidos/src/main/java/com/proyectosPersonales/springboot/app/gastos/dto/Deuda.java
@@ -34,7 +34,7 @@ public class Deuda {
 	@Column(name = "idDeuda")
 	private Integer idDeuda;
 
-	private Long importe; // importe del pago
+	private Double importe; // importe del pago
 
 	private String descripcion;// descripcion del pago
