+++ Misc/spring-boot-gastos-compartidos/src/main/java/com/proyectosPersonales/springboot/app/gastos/dto/UsuarioPago.java
@@ -13,7 +13,7 @@ public class UsuarioPago {
 	
 	private String codUsuario;
 	
-	private Long importe; 
+	private Double importe; 
 
 	private String descripcion;
